+++ app/channels/pairer/application_cable.rb
@@ -1,0 +1,9 @@
+module Pairer
+  module ApplicationCable
+    class Channel < ActionCable::Channel::Base
+    end
+
+    class Connection < ActionCable::Connection::Base
+    end
+  end
+end